Move 'utils' folder to parent directory for improved accessibility and organization
